+++ src/components/swap/FeesAndRateDetails.tsx
@@ -15,7 +15,11 @@ import { delayedFadeAnimation } from "../../animations/animations";
 import { SwapSavingsAndAddresses } from "./SwapSavingsAndAddresses";
 import { useSolanaWallet } from "../../hooks/useSolanaWallet";
 import { TooltipWrapper } from "../../common/ToolTipWrapper";
-import { formatAmount, formatAmountUsd } from "../../utils/utils";
+import {
+  formatAmount,
+  formatAmountUsd,
+  isAlpenSignetChain,
+} from "../../utils/utils";
 import { assetInfoStore } from "../../store/assetInfoStore";
 import { RateAndPriceDisplay } from "./RateAndPriceDisplay";
 
@@ -216,8 +220,16 @@ export const FeesAndRateDetails = () => {
       <AnimatePresence>
         {isDetailsExpanded && (
           <SwapSavingsAndAddresses
-            refundAddress={refundAddress}
-            receiveAddress={receiveAddress}
+            refundAddress={
+              inputAsset?.chain && isAlpenSignetChain(inputAsset.chain)
+                ? undefined
+                : refundAddress
+            }
+            receiveAddress={
+              outputAsset?.chain && isAlpenSignetChain(outputAsset.chain)
+                ? undefined
+                : receiveAddress
+            }
             showComparison={showComparisonHandler}
             networkFeesValue={Number(formatAmountUsd(networkFees, 0))}
           />