+++ src/utils/getTokenBalance.ts
@@ -1,4 +1,4 @@
-import { evmToViemChainMap, STARKNET_CONFIG } from "@gardenfi/core";
+import { evmToViemChainMap } from "@gardenfi/core";
 import {
   Asset,
   isBitcoin,
@@ -18,7 +18,7 @@ import {
 } from "viem";
 import { formatAmount } from "./utils";
 import { RpcProvider, Contract } from "starknet";
-import { network } from "../constants/constants";
+import { network, STARKNET_CONFIG } from "../constants/constants";
 import { Connection, PublicKey } from "@solana/web3.js";
 import logger from "./logger";
 import { getFullnodeUrl } from "@mysten/sui/client";