+++ src/utils/utils.ts
@@ -326,3 +326,14 @@ To learn more about Garden, refer to our documentation: https://docs.garden.fina
     );
   }
 };
+
+export const isAlpenSignetChain = (chain: string) => {
+  return chain.toLowerCase().includes("alpen_signet");
+};
+
+export const isStableCoinOrSeed = (asset: Asset) => {
+  return (
+    asset.symbol.toLowerCase().includes("usd") ||
+    asset.symbol.toLowerCase().includes("seed")
+  );
+};